feat!: find by date added
Changed the date range syntax to be more similar to author, i.e. instead
of doing `date-date`, you now specify `before:date,after:date`. Also
means we can use the tokenizer to parse the different parts of it, and
it will generalize better to when we do multiple flags as `or`.